requirements.txt
hyperliquid-python-sdk>=0.9.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
torch>=2.1.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
orjson>=3.9.0
pyarrow>=14.0.0
redis>=5.0.0
cryptography>=41.0.0
eth-account>=0.10.0
eth-utils>=2.2.0
msgpack>=1.0.0
optuna>=3.4.0
ta>=0.11.0
joblib>=1.3.0
matplotlib>=3.8.0
